Micro__Service/Micro_Service/requirements.txt
asgiref==3.4.1
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
setuptools==57.0.0
sqlparse==0.4.2
psycopg2-binary==2.8.6
PyJWT==2.3.0
wheel==0.36.2